+++ src/expander/quotes_removal.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   quotes_removal.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: atonkopi <atonkopi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jlabonde <jlabonde@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/25 18:34:58 by atonkopi          #+#    #+#             */
-/*   Updated: 2024/05/16 14:04:05 by atonkopi         ###   ########.fr       */
+/*   Updated: 2024/05/22 14:58:58 by jlabonde         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,7 +77,7 @@ void	remove_dollar_sign(t_token **token)
 void	remove_quotes(t_token *tokens)
 {
 	t_token	*temp;
-	char *unquoted_value;
+	char	*unquoted_value;
 
 	temp = tokens;
 	while (temp)
@@ -95,4 +95,3 @@ void	remove_quotes(t_token *tokens)
 		temp = temp->next;
 	}
 }
-
