+++ src/init.c
@@ -6,7 +6,7 @@
 /*   By: jlabonde <jlabonde@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/22 14:28:00 by atonkopi          #+#    #+#             */
-/*   Updated: 2024/05/22 14:37:18 by jlabonde         ###   ########.fr       */
+/*   Updated: 2024/05/22 14:59:27 by jlabonde         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,4 +27,3 @@ int	init_shell(t_shell *shell, char **env)
 	shell->heredoc = NULL;
 	return (EXIT_SUCCESS);
 }
-
